+++ demo/app/app.module.ts
@@ -32,6 +32,10 @@ import { PolygonPerformanceTestComponent } from './components/polygon-layer/poly
 import { KeyboardControlLayerComponent } from './components/keyboard-control-layer/keyboard-control-layer.component';
 import { LayerOrderComponent } from './components/layer-order/layer-order.component';
 import { AngularCesiumEntitiesEditorModule } from '../../src/angular-cesium-entities-editor/angular-cesium-entities-editor.module';
+import { EditorLayerComponent } from './components/editor-layer/editor-layer.component';
+import { SelectionLayerComponent } from './components/selection-layer/selection-layer.component';
+import { ContextMenuComponent } from './components/context-menu/context-menu.component';
+import { ContextMenuLayerComponent } from './components/context-menu-layer/context-menu-layer.component';
 import { PolygonsEditorLayerComponent } from './components/polygons-editor-layer/polygons-editor-layer.component';
 import { CirclesEditorLayerComponent } from './components/circles-editor-layer/circles-editor-layer.component';
 
@@ -62,6 +66,10 @@ import { CirclesEditorLayerComponent } from './components/circles-editor-layer/c
     KeyboardControlLayerComponent,
     HippodromeLayerComponent,
 		LayerOrderComponent,
+		EditorLayerComponent,
+    ContextMenuComponent,
+    ContextMenuLayerComponent,
+		SelectionLayerComponent,
 		PolygonsEditorLayerComponent,
 		CirclesEditorLayerComponent,
   ],
@@ -75,7 +83,7 @@ import { CirclesEditorLayerComponent } from './components/circles-editor-layer/c
     AppMaterialModule,
     ApolloModule.forRoot(getApolloClient),
   ],
-  entryComponents: [TracksDialogComponent],
+  entryComponents: [TracksDialogComponent, ContextMenuComponent],
   bootstrap: [AppComponent]
 })
 export class AppModule {