+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## NEXT
 ### Fixes
 * Service init order  [#131](https://github.com/TGFTech/angular-cesium/issues/131)
+* Added context menu service that allows dynamic injection of custom components.
+
+### Features
+* Added `selectionManagerService` and example in selection-layer.component.ts
+* Added context menu service that allows dynamic injection of custom components.
 * Added Circle editor `CirclesEditorService` - allow for creating circles and editing them from the map of from the code.
-* edited polygons are now draggable.
+* edited polygons are now draggable.  
 
 ## 0.0.29
 ### Features